backend/requirements.txt
# --- Core Flask Stack ---
Flask==3.1.2
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
MarkupSafe==3.0.3
click==8.3.0
blinker==1.9.0
colorama==0.4.6

# --- Flask Extensions ---
Flask-SQLAlchemy==3.1.1
Flask-Session==0.4.0
Flask-Cors==6.0.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1

# --- Security / Auth ---
bcrypt==5.0.0
PyJWT==2.10.1

# --- Database ---
SQLAlchemy==2.0.44
greenlet==3.2.4

# --- Pydantic / Typing ---
pydantic==1.10.12
typing_extensions==4.15.0

# --- AI / LLM ---
google-generativeai==0.8.3

# --- PDF Processing ---
pdfplumber==0.10.0
PyPDF2==3.0.1

# --- Utilities ---
python-dotenv==1.0.0
tqdm==4.65.0
Flask-RESTful==0.3.10